Relocating the event stream service.
