+++ src/services/husqvarna/eventStreamService.ts
@@ -1,11 +1,11 @@
-import { 
-    AutomowerEventStreamClient, AutomowerEvent, AutomowerEventTypes, ErrorEvent, 
-    PositionsEvent, SettingsEvent, StatusEvent 
-} from '../../../clients/automowerEventStreamClient';
-import { PlatformLogger } from '../../../diagnostics/platformLogger';
-import { BadCredentialsError } from '../../../errors/badCredentialsError';
-import { Timer } from '../../../primitives/timer';
-import { AccessTokenManager } from '../accessTokenManager';
+import {
+    AutomowerEvent, AutomowerEventStreamClient, AutomowerEventTypes, ErrorEvent,
+    PositionsEvent, SettingsEvent, StatusEvent
+} from '../../clients/automowerEventStreamClient';
+import { PlatformLogger } from '../../diagnostics/platformLogger';
+import { BadCredentialsError } from '../../errors/badCredentialsError';
+import { Timer } from '../../primitives/timer';
+import { AccessTokenManager } from './accessTokenManager';
 
 /**
  * A mechanism which is capable of streaming events for the Husqvarna account.
