+++ test/automowerPlatformSpy.ts
@@ -1,7 +1,7 @@
 import { AutomowerPlatform } from '../src/automowerPlatform';
 import { AccessTokenManager } from '../src/services/husqvarna/accessTokenManager';
-import { EventStreamService } from '../src/services/husqvarna/automower/eventStreamService';
 import { DiscoveryService } from '../src/services/husqvarna/discoveryService';
+import { EventStreamService } from '../src/services/husqvarna/eventStreamService';
 
 export class AutomowerPlatformSpy extends AutomowerPlatform {
     public discoveryService?: DiscoveryService;
